+++ src/Fauna/insfile_reader.cpp
@@ -272,14 +272,6 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
   }
   {
     const auto value =
-        find_hft_array_parameter<double>(table, "digestion.i_g_1992_ijk", true);
-    assert(value);
-    if (value->size() != 3)
-      throw bad_array_size("hf.digestion.i_g_1992_ijk", value->size(), "3");
-    for (int i = 0; i < 3; i++) hft.digestion_i_g_1992_ijk[i] = (*value)[i];
-  }
-  {
-    const auto value =
         find_hft_parameter<std::string>(table, "digestion.limit", true);
     assert(value);
     if (lowercase(*value) == lowercase("None"))
@@ -498,6 +490,16 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
     const auto value = find_hft_parameter<double>(
         table, "digestion.allometric.exponent", mandatory);
     if (value) hft.digestion_allometric.exponent = *value;
+  }
+  {
+    const bool mandatory =
+        (hft.digestion_limit == DigestiveLimit::IlliusGordon1992);
+    const auto value = find_hft_array_parameter<double>(
+        table, "digestion.i_g_1992_ijk", mandatory);
+    assert(value);
+    if (value->size() != 3)
+      throw bad_array_size("hft.digestion.i_g_1992_ijk", value->size(), "3");
+    for (int i = 0; i < 3; i++) hft.digestion_i_g_1992_ijk[i] = (*value)[i];
   }
   {
     const bool mandatory = (hft.expenditure_components.count(